refactor: separate code at configs(#208)
