+++ strawberry/src/data/config/BaseResponse.tsx
@@ -1,0 +1,7 @@
+interface BaseResponse<T> {
+  state: number;
+  message: string;
+  data: T;
+}
+
+export default BaseResponse;
