--- strawberry/src/data/config/request.tsx
@@ -1,9 +1,0 @@
-export type RequestMethod = "GET" | "POST" | "PUT" | "DELETE";
-
-export interface RequestConfig<T> {
-  url: string;
-  method: RequestMethod;
-  params?: Record<string, string>;
-  queryParams?: Record<string, string>;
-  body?: T;
-}
