--- strawberry/src/data/config/response.tsx
@@ -1,8 +1,0 @@
-import { CustomError } from "./customError";
-
-export interface FetchReturn<T> {
-  data: T | undefined;
-  loading: boolean;
-  statusCode: number;
-  error: CustomError | null;
-}
